230817 자신이 생성한 게시물을 삭제하기 까지
그러나 delete 오류 발생으로 넘어감

+++ src/boards/boards.service.ts
@@ -12,8 +12,14 @@ import { User } from 'src/auth/user.entity';
 export class BoardsService {
     constructor(private boardRepository: BoardRepository) { }
 
-    async getAllBoards(): Promise<Board[]> {
-        return this.boardRepository.find();
+    async getAllBoards(
+        user: User,
+    ): Promise<Board[]> {
+        const query = this.boardRepository.createQueryBuilder('board');
+        query.where('board.userId = :userId', { userId: user.id });
+        const boards = await query.getMany();
+
+        return boards;
     }
     // getAllBoards(): Board[] {
     //     return this.boards;
@@ -56,8 +62,23 @@ export class BoardsService {
     //     return found;
     // }
 
-    async deleteBoard(id: number): Promise<void> {
+    async deleteBoard(id: number, user: User): Promise<void> {
         const result = await this.boardRepository.delete(id);
+
+
+        const query = this.boardRepository.createQueryBuilder('board').delete()
+            .where('board.userId = :userId', { userId: user.id })
+            .andWhere('board.id = :id', { id });
+        //.where('board.id = :id', {id});
+        //const boards = await query.getMany();
+        //console.log(boards);
+        //const result = await query.execute();
+        // await this.boardRepository.createQueryBuilder('board')
+        //     //.leftJoin('board.userId', 'userId')
+        //     .where('board.userId = :userId', { userId: user.id })
+        //     .andWhere('board.id = :id', { id })
+        //     .delete()
+        //     .execute();
 
         if (result.affected === 0) {
             throw new NotFoundException(`Can't find Board with id ${id}`)
